+++ c++/pod5_format/c_api.cpp
@@ -739,6 +739,9 @@ pod5_error_t pod5_get_end_reason(
     case pod5::ReadEndReason::analysis_config_change:
         *end_reason_value = POD5_END_REASON_ANALYSIS_CONFIG_CHANGE;
         break;
+    case pod5::ReadEndReason::paused:
+        *end_reason_value = POD5_END_REASON_PAUSED;
+        break;
     default:
     case pod5::ReadEndReason::unknown:
         *end_reason_value = POD5_END_REASON_UNKNOWN;